more compatibility operators
