+++ brainpy/_src/math/operators/__init__.py
@@ -7,18 +7,15 @@ Operators for brain dynamics modeling.
 from . import (
   op_register,
   pre_syn_post,
-  wrap_jax,
   sparse_matmul,
 )
 
 __all__ = (
     op_register.__all__
     + pre_syn_post.__all__
-    + wrap_jax.__all__
     + sparse_matmul.__all__
 )
 
 from .sparse_matmul import *
 from .op_register import *
 from .pre_syn_post import *
-from .wrap_jax import *
