+++ brainpy/_src/math/operators/op_register.py
@@ -3,11 +3,11 @@
 import warnings
 from typing import Callable
 
-import brainpylib
 from jax.tree_util import tree_map
 
 from brainpy._src.math.object_transform.base import BrainPyObject
 from brainpy._src.math.ndarray import Array
+from brainpy._src.tools.package import import_brainpylib
 
 __all__ = [
   'XLACustomOp',
@@ -80,6 +80,7 @@ class XLACustomOp(BrainPyObject):
       gpu_func = None
 
     # register OP
+    brainpylib = import_brainpylib()
     self.op = brainpylib.register_op_with_numba(
       self.name,
       cpu_func=cpu_func,
